Add unit tests to UiTaskList & TaskList
