+++ src/test/java/vimification/model/LogicTaskListTest.java
@@ -5,32 +5,11 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.util.List;
-
 import org.junit.jupiter.api.Test;
 
 import vimification.model.task.Task;
 
 public class LogicTaskListTest {
-    @Test
-    public void testSimpleConstructor() {
-        LogicTaskList taskList = new TaskList();
-        assertTrue(taskList.isEmpty());
-        assertEquals(taskList.size(), 0);
-    }
-
-    @Test
-    public void testAdvancedConstructor() {
-        List<Task> tasks = List.of(new Task("Buy milk"), new Task("Do essay"));
-        LogicTaskList taskList = new TaskList(tasks);
-        assertFalse(taskList.isEmpty());
-        assertEquals(taskList.size(), 2);
-
-        for (int i = 0; i < tasks.size(); i++) {
-            assertEquals(taskList.get(i), tasks.get(i));
-        }
-    }
-
     @Test
     public void testGet() {
         LogicTaskList taskList = new TaskList();
